Slight Modifications made
Removed post-list and post-create components. Created Controllers backend folder which contains routes/ code. /routes folder now simply contains the export paths.
