+++ AnxietyKnot/backend/middleware/check-auth.js
@@ -1,9 +1,8 @@
-/// TO EXTRACT ANY FILES THAT ARE PART OF THE INCOMING REQUEST TO CHECK IF USER IS AUTHENTICATED OR NOT
-/// PARSE REQUESTS TO DETERMINE IF REQ BE ALLOWED OR DENIED
-// CHECK/VERIFY OUTGOING AND INCOMING AUTHENTICATION WITH HELP OF JSONWEBTOKEN
-/// FUNCTION EXE ON INCOMING REQUEST AND RUNS BEFORE ANY FINAL FUNCTIONS
-// DO NOT ADD MIDDLEWARE IN /routes/user.js, set it in posts.js
-
+/*TO EXTRACT ANY FILES THAT ARE PART OF THE INCOMING REQUEST TO CHECK IF USER IS AUTHENTICATED OR NOT
+ PARSE REQUESTS TO DETERMINE IF REQ BE ALLOWED OR DENIED
+ CHECK/VERIFY OUTGOING AND INCOMING AUTHENTICATION WITH HELP OF JSONWEBTOKEN
+ FUNCTION EXE ON INCOMING REQUEST AND RUNS BEFORE ANY FINAL FUNCTIONS
+DO NOT ADD MIDDLEWARE IN /routes/user.js, set it in posts.js */
 
 const jwt = require("jsonwebtoken");
