+++ AnxietyKnot/backend/models/user.js
@@ -3,13 +3,13 @@ const mongoose = require("mongoose");
 
 
 /*require npm install --save mongoose-unique-validator */
-//const uniqueValidator = require("mongoose-unique-validator");
+const uniqueValidator = require("mongoose-unique-validator");
 
 const userSchema = mongoose.Schema({
   email: { type: String, required: true, unique: true },
   password: { type: String, required: true }
 });
 
-//userSchema.plugin(uniqueValidator);
+userSchema.plugin(uniqueValidator);
 
 module.exports = mongoose.model("User", userSchema);
