+++ AnxietyKnot/src/app/app.module.ts
@@ -23,7 +23,6 @@ import { MatGridListModule } from '@angular/material/grid-list';
 import { MatMenuModule } from '@angular/material/menu';
 import { ResourceComponent } from './resource/resource.component';
 import { TrackerComponent } from './tracker/tracker.component';
-//import { EntryComponent } from './entry/entry.component';
 import { CommonModule } from '@angular/common';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 import { QuillModule } from 'ngx-quill';
@@ -69,7 +68,6 @@ import { ErrorComponent } from "./error/error.component";
     ErrorComponent
     //PostCreateComponent,
     //PostListComponent,
-      //EntryComponent,
   ],
   imports: [
     BrowserModule,
