+++ AnxietyKnot/src/app/auth/login/login.component.ts
@@ -16,7 +16,7 @@ export class LoginComponent implements OnInit, OnDestroy {
   constructor(public authService: AuthService) {}
   ngOnInit() {
     this.authStatusSub = this.authService.getAuthStatusListener().subscribe(
-      authStatus => {
+      _authStatus => {
         this.isLoading = false;
       }
     );
