+++ AnxietyKnot/src/app/auth/signup/signup.component.ts
@@ -15,7 +15,7 @@ export class SignupComponent implements OnInit, OnDestroy {
   constructor(public authService: AuthService) {}
   ngOnInit() {
     this.authStatusSub = this.authService.getAuthStatusListener().subscribe(
-      authStatus => {
+      _authStatus => {
         this.isLoading = false;
       }
     );
