+++ AnxietyKnot/src/app/error/error.component.ts
@@ -1,9 +1,9 @@
 import { Component, Inject } from "@angular/core";
 
 import { MAT_DIALOG_DATA } from "@angular/material/dialog";
-// import { Subscription } from "rxjs";
+ import { Subscription } from "rxjs";
 
-// import { ErrorService } from "./error.service";
+ import { ErrorService } from "./error.service";
 
 @Component({
   templateUrl: "./error.component.html",
@@ -11,5 +11,22 @@ import { MAT_DIALOG_DATA } from "@angular/material/dialog";
   // styleUrls: ["./error.component.css"]
 })
 export class ErrorComponent {
-  constructor(@Inject(MAT_DIALOG_DATA) public data: { message: string }) {}
+  //data: { message: string };
+ private errorSub: Subscription;
+  constructor(@Inject(MAT_DIALOG_DATA) public data: { message: string }, private errorService: ErrorService) {}
+  // constructor(private errorService: ErrorService) {}
+
+  ngOnInit() {
+    this.errorSub = this.errorService.getErrorListener().subscribe(message => {
+     this.data = { message: message };
+     });
+   }
+
+  // onHandleError() {
+  //   this.errorService.handleError();
+  // }
+
+  // ngOnDestroy() {
+  //   this.errorSub.unsubscribe();
+  // }
 }
