+++ AnxietyKnot/src/app/error/error.service.ts
@@ -1,0 +1,19 @@
+import { Subject } from "rxjs";
+import { Injectable } from "@angular/core";
+
+@Injectable({ providedIn: "root" })
+export class ErrorService {
+  private errorListener = new Subject<string>();
+
+  getErrorListener() {
+    return this.errorListener.asObservable();
+  }
+
+  throwError(message: string) {
+    this.errorListener.next(message);
+  }
+
+  handleError() {
+    this.errorListener.next(null);
+  }
+}
