--- AnxietyKnot/src/app/post-create/post-create.component.css
@@ -1,21 +1,0 @@
-/*  you can also check out the other supported theme - bubble - with floating tooltips by including the css :
-cdn.quilljs.com/1.0.4/quill.bubble.css
- */
-
-@import "~quill/dist/quill.core.css";
-@import "~quill/dist/quill.bubble.css";
-@import "~quill/dist/quill.snow.css";
-
- body {
-  background-color: #222;
-  padding: 20px;
-}
-
-.container {
-  background-color: whitesmoke;
-}
-
-#quillEditor {
-  height: 180px;
-  overflow: auto;
-}
