--- AnxietyKnot/src/app/post-create/post-create.component.html
@@ -1,35 +1,0 @@
-<mat-card>
-  <form (submit)="onSavePost(postForm)" #postForm="ngForm">
-    <mat-form-field>
-      <input
-        matInput
-        type="text"
-        name="title"
-        [ngModel]="post?.title"
-        required
-        minlength="3"
-        placeholder="Journal Title"
-        #title="ngModel">
-      <mat-error *ngIf="title.invalid">Please enter a journal title.</mat-error>
-    </mat-form-field>
-    <mat-form-field>
-      <textarea
-        id="quillEditor"
-        matInput
-        rows="4"
-        name ="content"
-        [ngModel]="post?.content"
-        required
-        #content="ngModel"></textarea>
-      <mat-error *ngIf="content.invalid">Please enter your journal content!.</mat-error>
-    </mat-form-field>
-    <button
-      mat-raised-button
-      color="accent"
-      type="submit">Save Entry</button>
-  </form>
-</mat-card>
-
-
-
-
