--- AnxietyKnot/src/app/post-create/post-create.component.ts
@@ -1,62 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-import { NgForm } from "@angular/forms";
-import { ActivatedRoute, ParamMap } from "@angular/router";
-import Quill from "quill";
-import { Post } from "../post.model";
-
-import { PostsService } from "../posts.service";
-
-@Component({
-  selector: "app-post-create",
-  templateUrl: "./post-create.component.html",
-  styleUrls: ["./post-create.component.css"]
-})
-export class PostCreateComponent implements OnInit{
-  enteredTitle = "";
-  enteredContent = "";
-  private mode = 'create';
-  private postId: string;
-  public post: Post;
-  public static text: string;
-
-  constructor(public postsService: PostsService, public route: ActivatedRoute) {}
-
-  ngOnInit() {
-    this.route.paramMap.subscribe((paramMap: ParamMap) => {
-      if (paramMap.has('postId')) {
-        this.mode = 'edit';
-        this.postId = paramMap.get('postId');
-        this.post = this.postsService.getPost(this.postId);
-      } else {
-        this.mode = 'create';
-        this.postId = null;
-      }
-    });
-
-    var options = {
-      placeholder: 'Waiting for your precious content',
-      theme: 'snow'
-    };
-
-    var editor = new Quill('#quillEditor', options);
-
-    editor.on('text-change', function() {
-      PostCreateComponent.text = editor.getText() + "";
-    });
-  }
-
-  onSavePost(form: NgForm) {
-    if (form.invalid) {
-      return;
-    }
-    if (this.mode === 'create') {
-      this.postsService.addPost(form.value.title, form.value.content);
-    }
-    else {
-      this.postsService.updatePost(this.postId, form.value.title, form.value.content);
-    }
-    form.resetForm();
-  }
-
-}
-
