--- AnxietyKnot/src/app/post-list/post-list.component.css
@@ -1,3 +1,0 @@
-.grid-container {
-  margin: 20px;
-}
