--- AnxietyKnot/src/app/post-list/post-list.component.html
@@ -1,15 +1,0 @@
-<div class="grid-container">
-<mat-accordion multi="true" *ngIf="posts.length > 0">
-  <mat-expansion-panel *ngFor="let post of posts">
-    <mat-expansion-panel-header>
-      {{ post.title }}
-    </mat-expansion-panel-header>
-    <p>{{ replace(post.content) }}</p>
-    <mat-action-row>
-      <button mat-button color="primary" [routerLink]="['/edit', post.id]">EDIT</button>
-      <button mat-button color="warn" (click)="onDelete(post.id)">DELETE</button>
-    </mat-action-row>
-  </mat-expansion-panel>
-</mat-accordion>
-<p class="info-text mat-body-1" *ngIf="posts.length <= 0">No journal entries added yet!</p>
-</div>
