--- AnxietyKnot/src/app/post-list/post-list.component.ts
@@ -1,44 +1,0 @@
-import { Component, OnInit, OnDestroy } from "@angular/core";
-import { Subscription } from 'rxjs';
-
-import { Post } from "../post.model";
-import { PostsService } from "../posts.service";
-
-@Component({
-  selector: "app-post-list",
-  templateUrl: "./post-list.component.html",
-  styleUrls: ["./post-list.component.css"]
-})
-export class PostListComponent implements OnInit, OnDestroy {
-  // posts = [
-  //   { title: "First Post", content: "This is the first post's content" },
-  //   { title: "Second Post", content: "This is the second post's content" },
-  //   { title: "Third Post", content: "This is the third post's content" }
-  // ];
-  posts: Post[] = [];
-  public noHtmlContent: string[] = [];
-  private postsSub: Subscription = new Subscription;
-
-  constructor(public postsService: PostsService) {}
-
-  ngOnInit() {
-    this.postsService.getPosts();
-    this.postsSub = this.postsService.getPostUpdateListener()
-      .subscribe((posts: Post[]) => {
-        this.posts = posts;
-      });
-    }
-
-  replace(content: any) {
-    var parsedContent = content.replace(/<[^>]+>/g, '');
-    return parsedContent;
-  }
-
-  onDelete(postId: string) {
-    this.postsService.deletePost(postId);
-  }
-
-  ngOnDestroy() {
-    this.postsSub.unsubscribe();
-  }
-}
