+++ AnxietyKnot/src/app/prompted-entry.ts
@@ -17,3 +17,18 @@ export interface PromptedEntry {
   thinking_differently: string;
   creator: string;
   }
+/*
+  id: string;
+  title: string;
+  what_happened: string;
+  going_through_mind: string;
+  emotion1: string;
+  intensity1: number;
+  emotion2: string;
+  intensity2: number;
+
+  thought_patterns: Array<string>;
+  custom_thought_patterns: string;
+  thinking_differently: string;
+  creator: string;
+*/
